+++ z3/src/ast/mod.rs
@@ -22,22 +22,18 @@ mod bv;
 mod array;
 mod set;
 mod seq;
+mod datatype;
 
 pub use array::Array;
 pub use bool::Bool;
 pub use bv::BV;
+pub use datatype::Datatype;
 pub use float::Float;
 pub use int::Int;
 pub use real::Real;
 pub use set::Set;
 pub use seq::Seq;
 pub use string::String;
-
-/// [`Ast`] node representing a datatype or enumeration value.
-pub struct Datatype {
-    pub(crate) ctx: Context,
-    pub(crate) z3_ast: Z3_ast,
-}
 
 /// A dynamically typed [`Ast`] node.
 pub struct Dynamic {
@@ -693,32 +689,6 @@ impl Dynamic {
         match self.sort_kind() {
             SortKind::Datatype => Some(unsafe { Datatype::wrap(&self.ctx, self.z3_ast) }),
             _ => None,
-        }
-    }
-}
-
-impl Datatype {
-    pub fn new_const<S: Into<Symbol>>(ctx: &Context, name: S, sort: &Sort) -> Self {
-        assert_eq!(ctx, &sort.ctx);
-        assert_eq!(sort.kind(), SortKind::Datatype);
-
-        unsafe {
-            Self::wrap(ctx, {
-                Z3_mk_const(ctx.z3_ctx.0, name.into().as_z3_symbol(ctx), sort.z3_sort)
-            })
-        }
-    }
-
-    pub fn fresh_const(ctx: &Context, prefix: &str, sort: &Sort) -> Self {
-        assert_eq!(ctx, &sort.ctx);
-        assert_eq!(sort.kind(), SortKind::Datatype);
-
-        unsafe {
-            Self::wrap(ctx, {
-                let pp = CString::new(prefix).unwrap();
-                let p = pp.as_ptr();
-                Z3_mk_fresh_const(ctx.z3_ctx.0, p, sort.z3_sort)
-            })
         }
     }
 }
